move to IndexGroupList
