+++ src/components/IndexGroupsListItem.tsx
@@ -1,0 +1,12 @@
+import React from "react"
+import styled from "styled-components"
+
+interface Props {
+
+}
+
+
+
+export default function IndexGroupsListItem({ id, name, setGroup } :Props) {
+
+}
